separate `config.rs` -> `config/`
