+++ database/src/config/mod.rs
@@ -1,0 +1,10 @@
+//! TODO
+
+mod config;
+pub use config::Config;
+
+mod reader_threads;
+pub use reader_threads::ReaderThreads;
+
+mod sync_mode;
+pub use sync_mode::SyncMode;
